refactoring

+++ Modules/Currency/Http/Controllers/Admin/CurrencyController.php
@@ -10,32 +10,42 @@ use Modules\Currency\Http\Requests\CurrencyCreateRequest;
 use Modules\Currency\Http\Requests\CurrencyUpdateRequest;
 use Modules\Currency\Http\Resources\CurrencyResource;
 use Modules\Currency\Models\Currency;
+use Modules\Currency\Repositories\CurrencyRepository;
 use Modules\Currency\Services\CurrencyStorage;
 
 class CurrencyController extends Controller
 {
     public function __construct(
-        protected CurrencyStorage $currencyStorage
-    ){
-        $this->authorizeResource(Currency::class, 'currency');
-    }
+        protected CurrencyStorage $currencyStorage,
+        protected CurrencyRepository $currencyRepository
+    ){}
 
     public function store(CurrencyCreateRequest $request)
     {
+        $this->authorize('create', Currency::class);
+
         $currency = $this->currencyStorage->store(CurrencyDto::fromFormRequest($request));
 
         return new CurrencyResource($currency);
     }
 
-    public function update(Currency $currency, CurrencyUpdateRequest $request)
+    public function update(int $currency, CurrencyUpdateRequest $request)
     {
+        $currency = $this->currencyRepository->find($currency);
+
+        $this->authorize('update', $currency);
+
         $currency = $this->currencyStorage->update($currency, CurrencyDto::fromFormRequest($request));
 
         return new CurrencyResource($currency);
     }
 
-    public function destroy(Currency $currency)
+    public function destroy(int $currency)
     {
+        $currency = $this->currencyRepository->find($currency);
+
+        $this->authorize('delete', $currency);
+
         $this->currencyStorage->delete($currency);
 
         return response()->noContent();
